Controls, Character, Decks & Cards
The controls are defined as done towards providing suitable input, there will need to be more for controls towards providing suitable events.

The character has their scriptable object in which include their suitable stats in the scriptable objects. In addition the player model can set suitable values based on the character data in order to provide suitable values for the player based upon the player's chosen character.

The decks have the beginning of object pooling which depending on the enumeration deck type will pool a certain amount of objects depending on the Enum's integer = deck capacity array's integer.

For the card, there is a scriptable object for offence, defence and movement cards which provide their card type along with the character that can obtain the card, the type of rarity and their unique values, leading to creating offence, defence and movement cards.