Update on Player Controller, Turn Manager & State Manager
There has been an update for 3 of the implementations.
Player Controller - Added a new Method to return the value of the current mana once is the player's turn.
Turn Manager - Was able to call the manager to the state of one of the two players which will change their state from inactive to start.
State Manager - Inactive, Start & Deciding States are able to provide suitable implementations (currently) with some requiring more implementations and potential tweaks. However at the end the player can go from inactive to deciding along with choosing the movement and status card during the deciding state and confirm their action.