All Character Cards
This commit is only creating the data cards that are obtained at the start of the game for everyone to use in the game. This includes:
- 3 Movement Cards
- 4 Offence Cards
- 4 Defence Cards

This will be use for the deck pool to create these cards to be use in the game, which can allow me to make progress on using movement cards and moving in the game.