Adding Starting Cards & Testing Prefabs
Adding starting cards in the character data which later on will be use to for the object pooling to enable suitable offence, defence and movement cards into the suitable decks.

In addition, a test script is made to check if the prefabs function using github. If they do I don't need to worry about missing data. Otherwise I'll have to fix the issue before creating prefabs later on.