Functional Deck Pooling, New Card & Potential State Change
The decks are now able to pool suitable starting cards from their specific decks.
Original Idea for Deck Pooling:
- Singular Deck Pool for All Decks
- However couldn't convert the data
Alternative Idea to Function the Decks:
- Singular Deck Pool for Each Deck
- Capable of storing Data in Objects

There are new starting cards which were the common cards that can provide testing for object pooling to create empty cards to useable cards for certain actions.

There is a considerations to update the StateBase to have MonoBehaviour in order to collect components from other objects, allowing the state manager to have a singular responsibility.