Updated State Machine & Movement Deck Pile
There has been an update of the state base which has the component become a MonoBehaviour, meaning that the other states can reference other components, leading the state machine to have the singular responsibility of switching states rather than calling and invoking other component events.

In addition there is a deck pile which will be used to draw cards for the player to decide on, for this instance the type of dice they want to roll.

There was an issue that crashed the unity project being that the while loop that was used was infinite, leading me to reset the progress made on this commit over again.