Successful Drawing Cards
I have manage to find the issue of the crash which was I included the draw card which provided 3 starting cards that were calling the draw card that cause the infinite while loop to occur.

This has been fixed by testing the drawing cards with an update that was removed, this can lead me to call the the method from the deciding state to collect the selected cards.