Roll State Added, Improve Controller & Cancel Interface
The Roll State which allows the player to either roll the dice depending on the card they use or cancel and return back to the deciding state.

The controller had added into the player model the roll value that can be collected for the move state (which will be implemented in the future)

The cancel interface didn't have a event to be called, which now has changed to include an event arg to listen to the event once invoked.