Moving Directions
Multiple Pathways are added which the player can choose a path to follow if there they landed on the direction space.

If the direction restriction is the same type as the player's current direction then this will be excluded from the choices.

If there is only one path to chose then the state automatically choses that state for the player.