Testing Turn Manager
This commit is to test if once the player has ended their turn whether this be during the move state or picking state (future commit) will move onto the next player.

This was another issue where the project crashed, however instead of able to identify the issue, I cannot tell the main issue. This will lead me to save as it currently is to identify the problem.