Walk returns to normal
Walk was tested for the move state and now has their own values back to normal.

Start State cause a crash due to enabling the event again when exiting when it should be disabling.

And testing if ending the turn is reached from the turn manager to check if the turn manager caused the crashing or the inactive state.