Player can call the manager to end turn
The player can go from the start state to the inactive and is capable of getting the turn manager to end the turn.

However, I still cannot figure out the solution on ensuring that both players can have their turn from start to finish.

This will need to be revised towards finding a suitable solution as there as unique outcomes that cause the crash to occur.