Reaper's Abilities Function & Combat System fixed
There was an issue with the while loop being that the loop couldn't provide the if statements to operate, meaning that the implementation has to expand from the one use ability script to the inactive and move states to reference the opponent and the player being defeated from being inactive whilst in the last reapsort form.

Combat system was fixed that instantly set to true once one of the players decided to confirm their choice causing the other player unable to get their choice in combat.