Improved Last Reapsort & Added Mana Behaviour with some Comments
This commit provides the change on the last Reapsort and the mana behaviour.
Added:
- Controller can now change the value of the current mana by using cards which will decrease the value of the cards.

Updates:
- Last Reapsort is now able to provide encapsulation of the combat system's defender to get the defender's player controller to check IsAlive bool is false
- Deciding & Roll State provide changes towards operating the movement card's mana cost, deciding now has the condition of if the mana cost is above the player's current mana will prevent the player from changing from the deciding to roll state.
ToDo:
The mana behaviour needs to find the lowest mana cost value out of all of the drawn cards to see if the player cannot draw any cards, if the player cannot then the player cannot provide the suitable action. For instance, the player cannot attack or defend if the player's mana is below the card with the lowest mana cost.