Apply Status Cards & Target State
This commit, I provided status card with their own deck. The status cards can be selected in the deciding state in which changes the state to the target state, this then selects 4 players to be targeted for the effect and when the player is selected will be affected by the status card's effect with a enumeration and int cooldown.