Adding Buffs & Resetting Stats when starting turn
Buffs are added into the project, whilst all of the buffs are in the component, the currentBuffs script requires more implementation as only the invincible is there.

In addition regain mana in the controller has changed to reset stats which before activating the effects will now reset the mana to the max mana (as it was usually) along with resetting the multipliers to 1 (thrust, guard & roll).